update UI for createQuestionDialog and fixed magic numbers in playArea tests

+++ client/src/app/components/admin-question/admin-question.component.ts
@@ -1,5 +1,6 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
+import { Router } from '@angular/router';
 import { CreateQuestionDialogComponent } from '@app/components/create-question-dialog/create-question-dialog.component';
 import { QuestionsService } from '@app/services/questions.service';
 import { Question } from '@common/game';
@@ -19,6 +20,7 @@ export class AdminQuestionComponent {
     constructor(
         public dialog: MatDialog,
         private questionsService: QuestionsService,
+        private router: Router,
     ) {}
 
     openDialog(): void {
@@ -38,7 +40,9 @@ export class AdminQuestionComponent {
     }
 
     deleteQuestion(question: Question): void {
-        this.questionsService.deleteQuestion(question);
+        if (this.router.url === '/admin/questions') {
+            this.questionsService.deleteQuestion(question);
+        }
         this.deleteRequest.emit(this.index);
     }
 }
